# public/openapi.yml
openapi: 3.0.0
info:
  title: cuenti time API
  description: |
    API pública multi-tenant. Cada Bearer token pertenece a **una sola empresa**.
    Todas las consultas se filtran automáticamente por esa empresa; no es posible
    leer ni escribir datos de otra compañía.
  version: 1.1.0
servers:
  - url: http://192.168.1.230:7578/api/v1
    description: API v1 (cuenti time)
  - url: /api/v1
    description: Relativo al host actual
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Token `cuenti_…` emitido en Integraciones. Scopes `read` o `read,write`.
...
security:
  - bearerAuth: []
paths:
...
  /employees:
    get:
      summary: Listar empleados
      tags: [Empleados]
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [ACTIVE, INACTIVE], default: ACTIVE }
        - name: branchId
          in: query
          schema: { type: string }
          description: Debe pertenecer a la misma empresa del token
      responses:
        "200": { description: Lista de empleados de la empresa del token }
...
  /positions:
    get:
      summary: Listar cargos
      tags: [Catálogos]
      parameters:
        - name: active
          in: query
          schema:
            type: string
            enum: ["true", "false", "all"]
            default: "true"
      responses:
        "200": { description: Cargos de la empresa }
...
  /incident-types:
    get:
      summary: Tipos de novedad
      tags: [Novedades]
      parameters:
        - name: active
          in: query
          schema:
            type: string
            enum: ["true", "false", "all"]
            default: "true"
      responses:
        "200": { description: Tipos de novedad }
...
  /branches:
    get:
      summary: Listar sucursales
      tags: [Sucursales]
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [ACTIVE, INACTIVE], default: ACTIVE }
      responses:
        "200": { description: Lista de sucursales }
...